ooo: fix slow startup
On first soffice start with an empty profile, the process
immediatly exits with returncode 81 - this is confirmed in
https://bugs.documentfoundation.org/show_bug.cgi?id=107912 and
not really a bug. This bug suggest running with soffice, but
this seems not really needed for us, because we expect that
libreoffice may crash and we restart it in that case.

Problem is that util.waitStartDaemon(process) does not check
if process is still running, so if the process terminates we
have to wait for the timeout.

This changes waitStartDaemon to check for terminated process
in every iteration and to stop waiting when process is
terminated.

waitStopDaemon is also changed to make use of the new hasExited
method for consistency.