testOooServer: fix test converting invalid file
This test tried tries to convert a broken file, but this broken file was
"not broken enough" and it managed to pass and reply with 200 OK.

Adjust the test to use a "more broken file" that actually cause a 402
return code.

This is because with libreoffice 7.5.2.2 that file was broken enough
to trigger the error, but changing here minimize the changes related to
libreoffice version up.