wkhtmltopdf: Improve error reporting
When `wkhtmltopdf` failed, there was no error reporting, apart from the
fact that the output file was missing.

This commits modernizes the call to `wkhtmltopdf` to use the new `run`
function available in Python 3, which supports a `check` parameter to
check the value returned by the program, and raise an exception if it is
not 0, including all relevant information.
In addition to that, any possible message that `wkhtmltopdf` outputs
in a successful execution is now logged.